+++ src/dznpy/support_files/ilog.py
@@ -82,7 +82,7 @@ struct ILogWithContext : ILog
 def create_header(namespace_prefix: NameSpaceIds = None) -> GeneratedContent:
     """Create the c++ header file contents that facilitates strict port typing."""
 
-    ns, cpp_ns = initialize_ns(namespace_prefix)
+    ns, cpp_ns, file_ns = initialize_ns(namespace_prefix)
     header = CommentBlock([header_hh_template(cpp_ns),
                            BLANK_LINE,
                            TEXT_GEN_DO_NOT_MODIFY,
@@ -92,7 +92,7 @@ def create_header(namespace_prefix: NameSpaceIds = None) -> GeneratedContent:
     includes = SystemIncludes(['functional', 'string'])
     body = Namespace(ns, contents=TextBlock([BLANK_LINE, body_hh(), BLANK_LINE]))
 
-    return GeneratedContent(filename=f'{"_".join(ns)}_ILog.hh',
+    return GeneratedContent(filename=f'{file_ns}_ILog.hh',
                             contents=str(TextBlock([header,
                                                     BLANK_LINE,
                                                     includes,