+++ src/dznpy/support_files/mutex_wrapped.py
@@ -79,7 +79,7 @@ private:
 def create_header(namespace_prefix: NameSpaceIds = None) -> GeneratedContent:
     """Create the c++ header file contents that facilitates strict port typing."""
 
-    ns, cpp_ns = initialize_ns(namespace_prefix)
+    ns, cpp_ns, file_ns = initialize_ns(namespace_prefix)
     header = CommentBlock([header_hh_template(cpp_ns),
                            BLANK_LINE,
                            TEXT_GEN_DO_NOT_MODIFY,
@@ -89,7 +89,7 @@ def create_header(namespace_prefix: NameSpaceIds = None) -> GeneratedContent:
     includes = SystemIncludes(['memory', 'mutex'])
     body = Namespace(ns, contents=TextBlock([BLANK_LINE, body_hh(), BLANK_LINE]))
 
-    return GeneratedContent(filename=f'{"_".join(ns)}_MutexWrapped.hh',
+    return GeneratedContent(filename=f'{file_ns}_MutexWrapped.hh',
                             contents=str(TextBlock([header,
                                                     BLANK_LINE,
                                                     includes,