+++ src/dznpy/support_files/strict_port.py
@@ -86,7 +86,7 @@ void ConnectPorts(Mts<P> provided, Mts<P> required)
 def create_header(namespace_prefix: NameSpaceIds = None) -> GeneratedContent:
     """Create the c++ header file contents that facilitates strict port typing."""
 
-    ns, cpp_ns = initialize_ns(namespace_prefix)
+    ns, cpp_ns, file_ns = initialize_ns(namespace_prefix)
     header = CommentBlock([header_hh_template(cpp_ns),
                            BLANK_LINE,
                            TEXT_GEN_DO_NOT_MODIFY,
@@ -95,7 +95,7 @@ def create_header(namespace_prefix: NameSpaceIds = None) -> GeneratedContent:
                            ])
     body = Namespace(ns, contents=TextBlock([BLANK_LINE, body_hh(), BLANK_LINE]))
 
-    return GeneratedContent(filename=f'{"_".join(ns)}_StrictPort.hh',
+    return GeneratedContent(filename=f'{file_ns}_StrictPort.hh',
                             contents=str(TextBlock([header,
                                                     BLANK_LINE,
                                                     body,