+++ test/unit_tests/support_files/test_sf_mutex_wrapped.py
@@ -96,6 +96,7 @@ def test_create_default_namespaced():
     assert result.namespace == ['Dzn']
     assert result.filename == 'Dzn_MutexWrapped.hh'
     assert result.contents == DEFAULT_DZN_NS_HH
+    assert result.contents_hash == '89427af31ed0b60636e25c82109be64b'
     assert 'namespace Dzn {' in result.contents